name changes, positionsystem added

+++ src/game/systems/healthregistry.h
@@ -3,13 +3,14 @@
 #include "../components/health.h"
 #include "../entity.h"
 #include <cstdint>
-#include <unordered_map>
 #include <cmath>
+#include <unordered_map>
+
 
 class HealthSystem
 {
-    std::unordered_map<uint_fast64_t, std::shared_ptr<Health>> healths;
-    // std::unordered_map<uint_fast64_t, Health> healths;
+    std::unordered_map<uint32_t, std::shared_ptr<Health>> healths;
+    // std::unordered_map<uint32_t, Health> healths;
 
 public:
     void add_entity(const Entity* entity)
@@ -17,12 +18,12 @@ public:
         healths.insert(entity->get_id(), entity->get_component<Health>());
     }
 
-    void remove_health_component(uint_fast64_t entity_id)
+    void remove_health_component(uint32_t entity_id)
     {
         healths.erase(entity_id);
     }
 
-    void purge(std::vector<uint_fast64_t> entity_ids)
+    void purge(std::vector<uint32_t> entity_ids)
     {
         for (auto entity_id : entity_ids)
         {
@@ -30,7 +31,7 @@ public:
         }
     }
 
-    void add_or_deduce_health(uint_fast64_t entity_id, uint_fast64_t value)
+    void add_or_deduce_health(uint32_t entity_id, uint32_t value)
     {
         // // value can be negative
         auto health = healths.at(entity_id).get();
@@ -53,11 +54,11 @@ public:
         }
     }
 
-    uint_fast64_t get_current_health(uint_fast64_t entity_id)
+    uint32_t get_current_health(uint32_t entity_id)
     {
         return healths.at(entity_id).current_health_points;
     }
-    uint_fast64_t get_max_health(uint_fast64_t entity_id)
+    uint32_t get_max_health(uint32_t entity_id)
     {
         return healths.at(entity_id).max_health_points;
     }
